AMBARI-24355. Logfeeder: include log type in checkpoint file names.
