+++ ambari-logsearch/ambari-logsearch-logfeeder/src/main/java/org/apache/ambari/logfeeder/common/ConfigHandler.java
@@ -421,7 +421,7 @@ public class ConfigHandler implements InputConfigMonitor {
   }
 
   public void cleanCheckPointFiles() {
-    inputManager.cleanCheckPointFiles();
+    inputManager.getCheckpointHandler().cleanupCheckpoints();
   }
 
   public void logStats() {
