+++ ambari-logsearch/ambari-logsearch-logfeeder/src/main/java/org/apache/ambari/logfeeder/input/InputManagerImpl.java
@@ -22,23 +22,14 @@ import com.google.common.annotations.VisibleForTesting;
 import org.apache.ambari.logfeeder.conf.LogFeederProps;
 import org.apache.ambari.logfeeder.docker.DockerContainerRegistry;
 import org.apache.ambari.logfeeder.docker.DockerContainerRegistryMonitor;
-import org.apache.ambari.logfeeder.input.monitor.CheckpointCleanupMonitor;
+import org.apache.ambari.logfeeder.plugin.manager.CheckpointManager;
 import org.apache.ambari.logfeeder.plugin.common.MetricData;
 import org.apache.ambari.logfeeder.plugin.input.Input;
 import org.apache.ambari.logfeeder.plugin.manager.InputManager;
-import org.apache.ambari.logfeeder.util.FileUtil;
-import org.apache.ambari.logfeeder.util.LogFeederUtil;
-import org.apache.commons.io.filefilter.WildcardFileFilter;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.log4j.Logger;
-import org.apache.solr.common.util.Base64;
 
 import javax.inject.Inject;
-import java.io.EOFException;
 import java.io.File;
-import java.io.FileFilter;
-import java.io.IOException;
-import java.io.RandomAccessFile;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -46,21 +37,15 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.UUID;
 
 public class InputManagerImpl extends InputManager {
 
   private static final Logger LOG = Logger.getLogger(InputManagerImpl.class);
 
-  private static final String CHECKPOINT_SUBFOLDER_NAME = "logfeeder_checkpoints";
-
   private Map<String, List<Input>> inputs = new HashMap<>();
   private Set<Input> notReadyList = new HashSet<>();
 
   private boolean isDrain = false;
-
-  private String checkPointExtension;
-  private File checkPointFolderFile;
 
   private MetricData filesCountMetric = new MetricData("input.files.count", true);
 
@@ -71,6 +56,9 @@ public class InputManagerImpl extends InputManager {
 
   @Inject
   private LogFeederProps logFeederProps;
+
+  @Inject
+  private CheckpointManager checkpointHandler;
 
   public List<Input> getInputList(String serviceName) {
     return inputs.get(serviceName);
@@ -130,43 +118,11 @@ public class InputManagerImpl extends InputManager {
 
   @Override
   public void init() throws Exception {
-    initCheckPointSettings();
+    checkpointHandler.init(logFeederProps);
     startMonitorThread();
     startDockerMetadataThread();
   }
 
-  private void initCheckPointSettings() {
-    checkPointExtension = logFeederProps.getCheckPointExtension();
-    LOG.info("Determining valid checkpoint folder");
-    boolean isCheckPointFolderValid = false;
-    // We need to keep track of the files we are reading.
-    String checkPointFolder = logFeederProps.getCheckpointFolder();
-    if (!StringUtils.isEmpty(checkPointFolder)) {
-      checkPointFolderFile = new File(checkPointFolder);
-      isCheckPointFolderValid = verifyCheckPointFolder(checkPointFolderFile);
-    }
-
-    if (!isCheckPointFolderValid) {
-      // Let's use tmp folder
-      checkPointFolderFile = new File(logFeederProps.getTmpDir(), CHECKPOINT_SUBFOLDER_NAME);
-      LOG.info("Checking if tmp folder can be used for checkpoints. Folder=" + checkPointFolderFile);
-      isCheckPointFolderValid = verifyCheckPointFolder(checkPointFolderFile);
-      if (isCheckPointFolderValid) {
-        LOG.warn("Using tmp folder " + checkPointFolderFile + " to store check points. This is not recommended." +
-          "Please set logfeeder.checkpoint.folder property");
-      }
-    }
-
-    if (isCheckPointFolderValid) {
-      LOG.info("Using folder " + checkPointFolderFile + " for storing checkpoints");
-      // check checkpoint cleanup every 2000 min
-      Thread checkpointCleanupThread = new Thread(new CheckpointCleanupMonitor(this, 2000),"checkpoint_cleanup");
-      checkpointCleanupThread.setDaemon(true);
-      checkpointCleanupThread.start();
-    } else {
-      throw new IllegalStateException("Could not determine the checkpoint folder.");
-    }
-  }
 
   private void startDockerMetadataThread() {
     if (logFeederProps.isDockerContainerRegistryEnabled()) {
@@ -230,34 +186,6 @@ public class InputManagerImpl extends InputManager {
     }
   }
 
-  private boolean verifyCheckPointFolder(File folderPathFile) {
-    if (!folderPathFile.exists()) {
-      try {
-        if (!folderPathFile.mkdir()) {
-          LOG.warn("Error creating folder for check point. folder=" + folderPathFile);
-        }
-      } catch (Throwable t) {
-        LOG.warn("Error creating folder for check point. folder=" + folderPathFile, t);
-      }
-    }
-
-    if (folderPathFile.exists() && folderPathFile.isDirectory()) {
-      // Let's check whether we can create a file
-      File testFile = new File(folderPathFile, UUID.randomUUID().toString());
-      try {
-        testFile.createNewFile();
-        return testFile.delete();
-      } catch (IOException e) {
-        LOG.warn("Couldn't create test file in " + folderPathFile.getAbsolutePath() + " for checkPoint", e);
-      }
-    }
-    return false;
-  }
-
-  public File getCheckPointFolderFile() {
-    return checkPointFolderFile;
-  }
-
   @Override
   public void addToNotReady(Input notReadyInput) {
     notReadyList.add(notReadyInput);
@@ -283,100 +211,6 @@ public class InputManagerImpl extends InputManager {
 
     filesCountMetric.value = getActiveFilesCount();
     // TODO: logStatForMetric(filesCountMetric, "Stat: Files Monitored Count", "");
-  }
-
-
-  public void cleanCheckPointFiles() {
-    if (checkPointFolderFile == null) {
-      LOG.info("Will not clean checkPoint files. checkPointFolderFile=" + checkPointFolderFile);
-      return;
-    }
-    LOG.info("Cleaning checkPoint files. checkPointFolderFile=" + checkPointFolderFile.getAbsolutePath());
-    try {
-      // Loop over the check point files and if filePath is not present, then move to closed
-      String searchPath = "*" + checkPointExtension;
-      FileFilter fileFilter = new WildcardFileFilter(searchPath);
-      File[] checkPointFiles = checkPointFolderFile.listFiles(fileFilter);
-      int totalCheckFilesDeleted = 0;
-      for (File checkPointFile : checkPointFiles) {
-        if (checkCheckPointFile(checkPointFile)) {
-          totalCheckFilesDeleted++;
-        }
-      }
-      LOG.info("Deleted " + totalCheckFilesDeleted + " checkPoint file(s). checkPointFolderFile=" +
-        checkPointFolderFile.getAbsolutePath());
-
-    } catch (Throwable t) {
-      LOG.error("Error while cleaning checkPointFiles", t);
-    }
-  }
-
-  private boolean checkCheckPointFile(File checkPointFile) {
-    boolean deleted = false;
-    try (RandomAccessFile checkPointReader = new RandomAccessFile(checkPointFile, "r")) {
-      int contentSize = checkPointReader.readInt();
-      byte b[] = new byte[contentSize];
-      int readSize = checkPointReader.read(b, 0, contentSize);
-      if (readSize != contentSize) {
-        LOG.error("Couldn't read expected number of bytes from checkpoint file. expected=" + contentSize + ", read="
-          + readSize + ", checkPointFile=" + checkPointFile);
-      } else {
-        String jsonCheckPointStr = new String(b, 0, readSize);
-        Map<String, Object> jsonCheckPoint = LogFeederUtil.toJSONObject(jsonCheckPointStr);
-
-        String logFilePath = (String) jsonCheckPoint.get("file_path");
-        String logFileKey = (String) jsonCheckPoint.get("file_key");
-        Integer maxAgeMin = null;
-        if (jsonCheckPoint.containsKey("max_age_min")) {
-          maxAgeMin = Integer.parseInt(jsonCheckPoint.get("max_age_min").toString());
-        }
-        if (logFilePath != null && logFileKey != null) {
-          boolean deleteCheckPointFile = false;
-          File logFile = new File(logFilePath);
-          if (logFile.exists()) {
-            Object fileKeyObj = FileUtil.getFileKey(logFile);
-            String fileBase64 = Base64.byteArrayToBase64(fileKeyObj.toString().getBytes());
-            if (!logFileKey.equals(fileBase64)) {
-              LOG.info("CheckPoint clean: File key has changed. old=" + logFileKey + ", new=" + fileBase64 + ", filePath=" +
-                logFilePath + ", checkPointFile=" + checkPointFile.getAbsolutePath());
-              deleteCheckPointFile = !wasFileRenamed(logFile.getParentFile(), logFileKey);
-            } else if (maxAgeMin != null && maxAgeMin != 0 && FileUtil.isFileTooOld(logFile, maxAgeMin)) {
-              deleteCheckPointFile = true;
-              LOG.info("Checkpoint clean: File reached max age minutes (" + maxAgeMin + "):" + logFilePath);
-            }
-          } else {
-            LOG.info("CheckPoint clean: Log file doesn't exist. filePath=" + logFilePath + ", checkPointFile=" +
-              checkPointFile.getAbsolutePath());
-            deleteCheckPointFile = !wasFileRenamed(logFile.getParentFile(), logFileKey);
-          }
-          if (deleteCheckPointFile) {
-            LOG.info("Deleting CheckPoint file=" + checkPointFile.getAbsolutePath() + ", logFile=" + logFilePath);
-            checkPointFile.delete();
-            deleted = true;
-          }
-        }
-      }
-    } catch (EOFException eof) {
-      LOG.warn("Caught EOFException. Ignoring reading existing checkPoint file. " + checkPointFile);
-    } catch (Throwable t) {
-      LOG.error("Error while checking checkPoint file. " + checkPointFile, t);
-    }
-
-    return deleted;
-  }
-
-  private boolean wasFileRenamed(File folder, String searchFileBase64) {
-    for (File file : folder.listFiles()) {
-      Object fileKeyObj = FileUtil.getFileKey(file);
-      String fileBase64 = Base64.byteArrayToBase64(fileKeyObj.toString().getBytes());
-      if (searchFileBase64.equals(fileBase64)) {
-        // even though the file name in the checkpoint file is different from the one it was renamed to, checkpoint files are
-        // identified by their name, which is generated from the file key, which would be the same for the renamed file
-        LOG.info("CheckPoint clean: File key matches file " + file.getAbsolutePath() + ", it must have been renamed");
-        return true;
-      }
-    }
-    return false;
   }
 
   public void waitOnAllInputs() {
@@ -469,5 +303,11 @@ public class InputManagerImpl extends InputManager {
     return logFeederProps;
   }
 
+  public CheckpointManager getCheckpointHandler() {
+    return checkpointHandler;
+  }
 
+  public void setCheckpointHandler(CheckpointManager checkpointHandler) {
+    this.checkpointHandler = checkpointHandler;
+  }
 }
