+++ ambari-logsearch/ambari-logsearch-logfeeder/src/main/java/org/apache/ambari/logfeeder/input/monitor/CheckpointCleanupMonitor.java
@@ -18,7 +18,7 @@
  */
 package org.apache.ambari.logfeeder.input.monitor;
 
-import org.apache.ambari.logfeeder.plugin.manager.InputManager;
+import org.apache.ambari.logfeeder.plugin.manager.CheckpointManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -27,11 +27,11 @@ public class CheckpointCleanupMonitor implements Runnable {
   private static final Logger LOG = LoggerFactory.getLogger(CheckpointCleanupMonitor.class);
 
   private long waitIntervalMin;
-  private InputManager inputManager;
+  private CheckpointManager checkpointHandler;
 
-  public CheckpointCleanupMonitor(InputManager inputManager, long waitIntervalMin) {
+  public CheckpointCleanupMonitor(CheckpointManager checkpointHandler, long waitIntervalMin) {
     this.waitIntervalMin = waitIntervalMin;
-    this.inputManager = inputManager;
+    this.checkpointHandler = checkpointHandler;
   }
 
   @Override
@@ -39,7 +39,7 @@ public class CheckpointCleanupMonitor implements Runnable {
     while (!Thread.currentThread().isInterrupted()) {
       try {
         Thread.sleep(1000 * 60 * waitIntervalMin);
-        inputManager.cleanCheckPointFiles();
+        checkpointHandler.cleanupCheckpoints();
       } catch (Exception e) {
         LOG.error("Cleanup checkpoint files thread interrupted.", e);
       }
