+++ ambari-logsearch/ambari-logsearch-logfeeder/src/test/java/org/apache/ambari/logfeeder/input/InputFileTest.java
@@ -26,8 +26,10 @@ import java.util.List;
 
 import org.apache.ambari.logfeeder.conf.LogEntryCacheConfig;
 import org.apache.ambari.logfeeder.conf.LogFeederProps;
+import org.apache.ambari.logfeeder.input.file.checkpoint.FileCheckpointManager;
 import org.apache.ambari.logfeeder.plugin.filter.Filter;
 import org.apache.ambari.logfeeder.plugin.input.InputMarker;
+import org.apache.ambari.logfeeder.plugin.manager.CheckpointManager;
 import org.apache.ambari.logfeeder.plugin.manager.InputManager;
 import org.apache.ambari.logsearch.config.json.model.inputconfig.impl.InputFileDescriptorImpl;
 import org.apache.commons.io.FileUtils;
@@ -77,6 +79,13 @@ public class InputFileTest {
     FileUtils.cleanDirectory(TEST_DIR);
   }
 
+  @AfterClass
+  public static void deleteDir() throws IOException {
+    if (TEST_DIR.exists()) {
+      FileUtils.deleteDirectory(TEST_DIR);
+    }
+  }
+
   @Before
   public void setUp() throws Exception {
     logFeederProps = new LogFeederProps();
@@ -85,6 +94,7 @@ public class InputFileTest {
     logEntryCacheConfig.setCacheLastDedupEnabled(false);
     logEntryCacheConfig.setCacheSize(10);
     logFeederProps.setLogEntryCacheConfig(logEntryCacheConfig);
+    logFeederProps.setCheckpointFolder("process3_checkpoint");
     testInputMarker = new InputFileMarker(inputFile, "", 0);
   }
 
@@ -125,13 +135,13 @@ public class InputFileTest {
   public void testInputFile_process3Rows() throws Exception {
     LOG.info("testInputFile_process3Rows()");
 
-    File checkPointDir = createCheckpointDir("process3_checkpoint");
     File testFile = createFile("process3.log");
 
     init(testFile.getAbsolutePath());
 
     InputManager inputManager = EasyMock.createStrictMock(InputManager.class);
-    EasyMock.expect(inputManager.getCheckPointFolderFile()).andReturn(checkPointDir);
+    CheckpointManager checkpointManager = new FileCheckpointManager();
+    EasyMock.expect(inputManager.getCheckpointHandler()).andReturn(checkpointManager);
     EasyMock.replay(inputManager);
     inputFile.setInputManager(inputManager);
